{Add} Interactivity - JDMHelper - Analyse syntaxique {Update} Database

+++ Database/Entities/UserEntity.cs
@@ -17,7 +17,7 @@ namespace BotJDM.Database.Entities
 
         public string Username { get; set; } = string.Empty;
 
-        public int TrustFactor { get; set; } = 0;
+        public float TrustFactor { get; set; } = 0;
 
         public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
     }
